Меня только свою ветку

+++ src/main/java/ru/skypro/homework/repositories/CommentsForAdsRepository.java
@@ -1,7 +1,8 @@
 package ru.skypro.homework.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import ru.skypro.homework.dto.Comment;
+import ru.skypro.homework.entity.Comment;
 
-public interface CommentsForAdsRepository extends JpaRepository<Comment,Long> {
+
+public interface CommentsForAdsRepository extends JpaRepository<Comment,Integer> {
 }
